// frontend/src/app/models/registrationReq.ts
export class Registration{
    username: string;
    password: string;
    phone: string;
    email: string;
    first_name: string;
    last_name: string;
    name: string;
    address: string;
    idNumber: number;
    description: string;
    userType: string;
    image: string;
    state: string;
    id: number;
}